refactoring: move logic to standalone module
